some changes

+++ app/commands/drop_db.py
@@ -2,6 +2,7 @@ import logging
 from flask_script import Command
 
 from app import db
+from app.local_settings import PRODUCTION_ENVIRONMENT
 
 logger = logging.getLogger('drop_db')
 
@@ -15,7 +16,10 @@ class DropDbCommand(Command):
 
 def drop_db():
     """ Initialize the database."""
-    logger.info('Dropping Db.')
-    # Recreate the database
-    db.drop_all()
-    logger.info('Dropped Db.')
+    if not PRODUCTION_ENVIRONMENT:
+        logger.info('Dropping Db.')
+        # Recreate the database
+        db.drop_all()
+        logger.info('Dropped Db.')
+    else:
+        logger.info('Not dropping since we are in production.')
